@@ -1,4 +1,4 @@
 """ReMap Network Security Scanner."""
-from .__version__ import __version__, get_version_string
+from .version import __version__, get_version_string
 
 __all__ = ['__version__', 'get_version_string']
