@@ -1,5 +1,4 @@
-"""GUI components."""
+"""GUI components for the ReMap application."""
 from .main_window import MainWindow
-from .styles import configure_styles, ReMapTheme
 
-__all__ = ['MainWindow', 'configure_styles', 'ReMapTheme']
+__all__ = ['MainWindow']
