@@ -2,22 +2,14 @@
 
 from dataclasses import dataclass
 from typing import List, Optional
-import ipaddress
 
-@dataclass
+@dataclass(frozen=True) # Use frozen to make it hashable for use in sets
 class Target:
     """Represents a scan target with IP and optional ports."""
     
     ip_address: str
-    ports: Optional[List[int]] = None
+    ports: Optional[tuple] = None # Use tuple to make it hashable
     hostname: Optional[str] = None
-    
-    def __post_init__(self):
-        """Validate IP address format."""
-        try:
-            ipaddress.ip_address(self.ip_address)
-        except ValueError:
-            raise ValueError(f"Invalid IP address: {self.ip_address}")
     
     @property
     def has_specific_ports(self) -> bool:
@@ -25,6 +17,6 @@ class Target:
         return self.ports is not None and len(self.ports) > 0
     
     def __str__(self):
-        if self.has_specific_ports:
-            return f"{self.ip_address}:{','.join(map(str, self.ports))}"
+        if self.has_specific_ports and self.ports is not None:
+            return f"{self.ip_address}:{','.join(map(str, sorted(self.ports)))}"
         return self.ip_address
