@@ -1,7 +1,7 @@
-"""Utility modules."""
+"""Utility modules for configuration, logging, validation, and file handling."""
 from .logger import setup_logger
 from .config import ConfigManager
-from .validators import Validators
+from .validators import Validators, ValidationError
 from .file_handler import FileHandler
 
-__all__ = ['setup_logger', 'ConfigManager', 'Validators', 'FileHandler']
+__all__ = ['setup_logger', 'ConfigManager', 'Validators', 'ValidationError', 'FileHandler']
